+++ src/main/java/co/nvqa/common/core/cucumber/glue/ApiOrderSteps.java
@@ -19,6 +19,7 @@ import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import io.restassured.response.Response;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import javax.inject.Inject;
@@ -378,7 +379,7 @@ public class ApiOrderSteps extends CoreStandardSteps {
       getOrderClient().updatePriorityLevelOfTransaction(orderId, priorityLevel.intValue());
     }, "API Core - Update priority level of an order");
   }
-
+  
   /**
    * @param dataTableAsMap <br><b>trackingId:</b>
    *                     {KEY_LIST_OF_CREATED_TRACKING_IDS[1]}<br><b>comment:</b> test comment
@@ -395,4 +396,27 @@ public class ApiOrderSteps extends CoreStandardSteps {
     }, "API Core - Operator post Lazada 3PL");
   }
 
+  @Given("API Core -  Wait for following order state:")
+  public void apiOperatorWaitForOrderStatus(Map<String, String> dataTableRaw) {
+    final Map<String, String> dataTable = resolveKeyValues(dataTableRaw);
+    Order expectedState = new Order();
+    expectedState.fromMap(dataTable);
+    int timeout = Integer.parseInt(dataTable.getOrDefault("timeout", "30"));
+    Assertions.assertThat(getOrderClient().waitUntilOrderState(expectedState, timeout, 1000))
+        .as("Order " + expectedState.getTrackingId() + " didn't get expected state " + dataTable)
+        .isTrue();
+  }
+
+  @Given("API Core - Verifies order state:")
+  public void apiOperatorVerifiesOrderState(Map<String, String> dataTableRaw) {
+    Map<String, String> dataTable = new HashMap<>(dataTableRaw);
+    dataTable.put("timeout", "1");
+    apiOperatorWaitForOrderStatus(dataTable);
+  }
+
+  @Given("API Core - wait for order state:")
+  public void waitForOrderState(Map<String, String> data) {
+    var expected = new Order(resolveKeyValues(data));
+    orderClient.waitUntilOrderState(expected, 5 * 60, 5000);
+  }
 }